update/ move getSearchStore to search file & move type to constant
